Change initial node list loading to happen after namespace adjustment
